Very small updates to PT Sensor procedure and driver code. Trying to make it easier to understand bugs from Arduino errors by printing exactly what is sent over the USB interface before trying to parse it with JSON

+++ source/procedures/PT_Sensor.py
@@ -57,7 +57,7 @@ def runPTSensor(arduino_reference, totalSensingTime, delayBetweenMeasurements, s
 	
 	while((time.time() - startTime) < totalSensingTime):
 		# Send a progress message
-		pipes.progressUpdate(share, 'Impedance Point', start=0, current=index, end=index * totalSensingTime / (time.time() - startTime), enableAbort=True)
+		pipes.progressUpdate(share, 'Impedance Point', start=0, current=index, end=int(index * totalSensingTime / (time.time() - startTime)), enableAbort=True)
 		index += 1
 
 		# If delayBetweenMeasurements is non-zero, wait before taking the measurement
